Fix a regression after 4cb2e0276 (Import time-related code from DDNet)
Commit aede6df2ca3b86b5b8c197731f0ca9d3c63031f5 in DDNet introduced a
time cache. The commit adds the cache invalidation on every new loop in
CServer::Run() while(m_RunServer).

In commit 4cb2e0276 the cache was added to the system.c but the cache
invalidation was missing.

It leaded to various side-effects because all calls to time_get() got
invalid results.